ok moving on to the next feature: as image is clicked or loaded and the animation starts and add validation two images shpuld be taken or uploaded and than animation starts
Next Need a functionality where it will check the both of the images  and compare both images and list whats missing from the image. It should display the name of objects missing between two images and the animation should run until it finds the missing objects and the name of missing objects should be shown in an popup with bold font.