fix : Please the analysing should be very accurate each and every item which is there in first image but not in second should be listed.
for example in image one there is a toy which is bit far and not completely visible , but in image 2 the place is empty it means the yellow toy is removed. It should be very accurate while analysing. Below is the image where at right side corner shelf there is a yellow toy but in image 2 its removed so that to identify.